+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## v0.1.0-beta.2 (2023/07/24)
+- Added api.getVersion()
+- Added static variables
+- Improved UI
+  - Added buttons to show/hide UI elements
+  - Added button to reload managers
 ## v0.1.0-beta.1 (2023/07/23)
 - Implemented ObjectViewer
   - load/reload `ObjectParameters/*.json`